Backend/requirements.txt
Pillow==7.2.0
my-voice-analysis==0.7
pycountry==20.7.3
django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
pytz==2020.1
faker==4.1.2
psycopg2-binary==2.8.5
boto3==1.14.44
gunicorn==20.0.4
gevent==20.9.0
celery==4.4.7
redis==3.5.3
praat-parselmouth==0.3.3
stripe==2.51.0
django-environ==0.4.5
social-auth-app-django==4.0.0